Preliminary support for launching software items
This is very preliminary support for launching software items.  This is gnarly because we are launching the software from within MAME and we need to do the work to identify all of the specific slots and the pertinent software parts.  Furthermore, this causes MAME to "cycle" because it causes a hard reset.

Caveats:
1.  This is not well tested at all
2.  For some unknown reason, when the software actually runs, subsequent worker_ui commands seem to be ignored.  This is likely a bug in the LUA code
3.  The overall experience with software items is pretty ugly, as when we add a software item to a folder, we provide an ugly context menu that lets the user choose any of the machines that might be relevant for this machine.  This experience needs to be refined