Partial fix to the MAME frozen search bar issue
The change I made ensures that we don't try to set the focus to the child window to which we attach MAME.  However, the search bar still seems to be frozen unless the user clicks away from BletchMAME and gives it focus back.  I do not yet understand this behavior.