Fixing the BletchMAME icon
The icon stopped displaying on windows.  It seems that Slint now has a way to specify an icon on a Window, and with this change BletchMAME now uses the technique supported by Slint.

Part of this change also involved changing the icon artifact from ICO to PNG.

A side effect of this is that the icon now shown on the Linux build.