Fixed an issue with slots that got ignored when building InfoDB
This was first reported as a crash on the "Devices & Images" dialog with `at486`

The issue was that we were ignoring "child devices" when setting up the InfoDB, so that we would ignore slots in `-listxml` that only existed because of the default Machine Config (e.g. - `ext:fdc:wd17xx:0` on `coco2b`).  Some drivers (like `at486`) seem to have child devices (i.e. - with a colon in the tag) that are not associated with a configurable device.

The fix was to stop ignoring these when building InfoDB, and to identify them when we create our MachineConfigs.