Overhaul of the BletchMAME application's state management
This is a general overhaul of how the BletchMAME application's overall state is handled.  BletchMAME does a bit of orchestration of MAME, both to run the actual emulation and also to get machine info with `-listxml`, and these interactions were largely handled on an ad hoc basis.  This is a brittle approach that could cause bad things to happen when unexpected conditions occur, and handling things like shutdowns and build skew were awkward.

This change attempts to centralize this state management, which should lay the foundation for a more resiliant application.  From a user facing perspective, the biggest change is that we show when MAME is resetting with a spinner, and should be able to prompt the user to retry if any of this fails.

Overall these changes have not been thoroughly tested.  Following are known issues
* Changing the paths can cause BletchMAME to hang on "Resetting MAME"
* When the emulation is shut down, the user can very briefly see the MAME main menu poking through
* If MAME dies unexpectedly (e.g. - missing ROMs); we really should present the user with an error similar to what would happen if a Machine Info reset fails