When BletchMAME is misconfigured, buttons for fixing paths on the main window will launch the Paths Dialog
Previously these would bring up a file dialog.  This works well for setting the MAME executable, but wasn't really good when setting the Plugins dialog, which requires specifying multiple paths.

Unfortunately when correcting a misconfiguration of the Plugins directory, the Paths Dialog will say "MAME Executable" even if it is showing the Plugins paths.  This seems to be a Slint bug:  https://github.com/slint-ui/slint/issues/7632