Partial fix to issue where we failed to report MAME failures
Including but not limited to "ROMs not found"

However, this is not fixed from the user's perspective, as not when you launch a machine where the ROM does not exist you get a blank screen.  There seems to be another issue pertaining to not being able to report messages to the user.  This same issue seems to cause the "Resetting MAME..." message that accompanies the spinner to not display.