Avoid unnecessary collection refreshes
Changes that affect history items (e.g. - changing search bar text) were causing the entire collection to be refreshed, which could be tens of thousands of items.  This change avoids this repopulation when the collection is not changed.

If this introduces a bug, it will take the form of something out there failing to refresh.