Created an initial (i.e. - incomplete) reimplementation of the Inputs dialog
This is in NO WAY done yet; more or less all we do at this point is acknowledge the input and input device related information from status updates, and build a non functional dialog vaguely resembling what the old BletchMAME had.  The code for interpreting input tokens is recognizable but not yet complete, and we are not yet doing anything special for analog inputs.

This is being checked in just because the foundational work was somewhat large, and even if it is non functional this represents a bit of work.  Just digesting the status updates is not trivial.