Updated to a newer version of our hacked fork of Slint and backend refactor
The Slint `winit` backend now uses a newer version of `winit`, which doesn't allow creation of windows outside of a context with an `ActiveEventLoop`.  This meant that the infrastructure for child window handling had to be overhauled.  Fortunately there is now a `CustomApplicationHandler` that provides this access.

This will be an improvement in the long run, and is necessary so that we can intercept keys like `ScrLk` to toggle the menu bar.

However this is an intrusive change, and even if it doesn't expose any new functionality there is a higher likelihood of regressions.