Added a preflight step to be ran prior to starting an emulation
This preflight step currently does nothing more than checking to see if the pertinent images exist, but this is a good start.  In the future we should also run our own ROM audit separate from MAME.

This also involved creating an `ImageDesc` object to represent an "image descriptor", to hide the details of dissecting image details.  There is likely going to be more work here.