Changing how we track "image descriptions" across BletchMAME
Traditionally, we would reference images using strings behaving the same as descriptors used on the MAME command line.  These could be
- Files (e.g. - `C:\foo\bar.dsk`)
- Software Names (e.g. - `zelda`)
- Sockets (e.g. - `socket.localhost:8888`)

The problem with using these descriptions is that while MAME will do its thing, BletchMAME doesn't really know what these are.  This means that if we have the word `zelda`, it could be either a software name or a file name.  Meaning that if we're going to preflight MAME emulation runs, we need to guess whether it is appropriate to do a file existence check.

With this change, the type of image is now encoded as an `ImageDesc`, which is now used across the codebase.

This was an intrusive change needing many changes (e.g. - encapsulating emulation start up arguments as `MameStartArgs`) affecting much of the codebase.  Therefore there is a high possibility of regressions.