Fixing an issue that could cause a panic if some folders have more items than others
The issue was that when we were responding to `Action::Configure`, we have to search for the folder.  But the code that looked up the item in the folder always executed even if the target folder was not found.  This would panic if the index of the configured item exceeded the size of the other folder.